Refactor(header): align text with content
