+++ src/components/pages/index.ts
@@ -1,2 +1,2 @@
 // page 컴포넌트를 모아놓은 파일
-export { default as Test } from "./Test";
+export { default as LocalCouncilReport } from "./LocalCouncilReport";
